- Updated the Readme - Improved the normal view shader - Added an option to display normals with or without the effect of normal mapping - Fixed more warnings

+++ DirectX11Engine/Shaders/SimplePixelShader.hlsl
@@ -39,7 +39,7 @@ cbuffer cbPerFrame
     DirectionalLight Sun;
     PointLight Lights[5];
     float3 CamPosition;
-    float LightsCount;
+    int LightsCount;
 };
 
 cbuffer cbPerObject
